Formatting

+++ src/AutoJIT.Runtime/Variant.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections;
-using System.Data.Odbc;
 using System.Globalization;
 using System.Text;
 using AutoJITRuntime.Exceptions;
@@ -10,59 +9,87 @@ namespace AutoJITRuntime
 {
     public abstract class Variant : IEnumerable
     {
-        protected abstract DataType DataType { get; }
+        protected abstract DataType DataType {
+            get;
+        }
 
         public virtual bool IsInt32 {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsInt64 {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsDouble {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsBool {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsString {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsArray {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsPtr {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsStruct {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsBinary {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsDefault {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual bool IsNull {
-            get { return false; }
+            get {
+                return false;
+            }
         }
 
         public virtual Variant this[ params int[] index ] {
-            get { throw new AutoJITRuntimerException( "" ); }
-            set { throw new AutoJITRuntimerException( "" ); }
+            get {
+                throw new AutoJITRuntimerException( "" );
+            }
+            set {
+                throw new AutoJITRuntimerException( "" );
+            }
         }
 
         public virtual IEnumerator GetEnumerator() {
-            return new VariantEnumerator( (IEnumerable) GetValue() );
+            return new VariantEnumerator( (IEnumerable)GetValue() );
         }
 
         public static Variant Create( Int32 @int ) {
@@ -122,126 +149,110 @@ namespace AutoJITRuntime
         }
 
         public static Variant Create( object @object ) {
-            if ( @object == null ) {
+            if( @object == null ) {
                 return new NullVariant();
             }
 
             var bytes = @object as byte[];
-            if ( bytes != null ) {
+            if( bytes != null ) {
                 return Create( bytes );
             }
 
-            if ( @object is int ) {
+            if( @object is int ) {
                 return Create( (int)@object );
             }
 
-            if (@object is Int64)
-            {
-                return Create((Int64)@object);
+            if( @object is Int64 ) {
+                return Create( (Int64)@object );
             }
 
-            if (@object is double)
-            {
-                return Create((double)@object);
+            if( @object is double ) {
+                return Create( (double)@object );
             }
 
-            if (@object is bool)
-            {
-                return Create((bool)@object);
+            if( @object is bool ) {
+                return Create( (bool)@object );
             }
 
             var str = @object as string;
-            if (str != null)
-            {
-                return Create(str);
+            if( str != null ) {
+                return Create( str );
             }
 
-            if (@object is IntPtr)
-            {
-                return Create((IntPtr)@object);
+            if( @object is IntPtr ) {
+                return Create( (IntPtr)@object );
             }
 
             var runtimeStruct = @object as IRuntimeStruct;
-            if (runtimeStruct != null)
-            {
-                return Create(runtimeStruct);
+            if( runtimeStruct != null ) {
+                return Create( runtimeStruct );
             }
 
             var @default = @object as Default;
-            if (@default != null)
-            {
-                return Create(@default);
+            if( @default != null ) {
+                return Create( @default );
             }
 
-            if (@object is char)
-            {
-                return Create((char)@object);
+            if( @object is char ) {
+                return Create( (char)@object );
             }
 
             var stringBuilder = @object as StringBuilder;
-            if (stringBuilder != null)
-            {
-                return Create(stringBuilder);
+            if( stringBuilder != null ) {
+                return Create( stringBuilder );
             }
 
             var vararr = @object as Variant[];
-            if (vararr != null)
-            {
-                return Create(vararr);
+            if( vararr != null ) {
+                return Create( vararr );
             }
 
             var var2d = @object as Variant[,];
-            if (var2d != null)
-            {
-                return Create(var2d);
+            if( var2d != null ) {
+                return Create( var2d );
             }
 
             var var3d = @object as Variant[,,];
-            if (var3d != null)
-            {
-                return Create(var3d);
+            if( var3d != null ) {
+                return Create( var3d );
             }
 
             var variant = @object as Variant;
-            if (variant != null)
-            {
+            if( variant != null ) {
                 return variant;
             }
 
-            if (@object is byte)
-            {
+            if( @object is byte ) {
                 return Create( (byte)@object );
             }
 
-            if ( @object is UIntPtr ) {
-                return Create( unchecked( (IntPtr) (long) (ulong) (UIntPtr) @object ) );
+            if( @object is UIntPtr ) {
+                return Create( unchecked( (IntPtr)(long)(ulong)(UIntPtr)@object ) );
             }
 
-            if ( @object is Int16 ) {
-                return Create( (int) (Int16) @object );
+            if( @object is Int16 ) {
+                return Create( (Int16)@object );
             }
 
-            if (@object is UInt16)
-            {
-                return Create((int)(UInt16)@object);
+            if( @object is UInt16 ) {
+                return Create( (UInt16)@object );
             }
 
-            if (@object is UInt32) {
+            if( @object is UInt32 ) {
                 var uint32 = (UInt32)@object;
-                if ( uint32 > int.MaxValue ||
-                     uint32 < int.MinValue ) {
-                    return Create( (Int64) uint32 );
+                if( uint32 > int.MaxValue
+                    || uint32 < int.MinValue ) {
+                    return Create( uint32 );
                 }
-                return Create( (Int32) uint32 );
+                return Create( (Int32)uint32 );
             }
 
-            if (@object is UInt64)
-            {
-                return Create((Int64)(UInt64)@object);
+            if( @object is UInt64 ) {
+                return Create( (Int64)(UInt64)@object );
             }
 
-            if ( @object is float ) {
-                return Create( (double) (float) @object );
+            if( @object is float ) {
+                return Create( (float)@object );
             }
             throw new NotImplementedException();
         }
@@ -275,7 +286,7 @@ namespace AutoJITRuntime
         public abstract Type GetRealType();
 
         public static Variant operator +( Variant a, Variant b ) {
-            switch (a.DataType) {
+            switch(a.DataType) {
                 case DataType.Null:
                 case DataType.Int32:
                 case DataType.Double:
@@ -290,7 +301,7 @@ namespace AutoJITRuntime
         }
 
         public static Variant operator -( Variant a ) {
-            switch (a.DataType) {
+            switch(a.DataType) {
                 case DataType.Int32:
                     return -a.GetInt();
                 case DataType.Int64:
@@ -303,7 +314,7 @@ namespace AutoJITRuntime
         }
 
         public static Variant operator -( Variant a, Variant b ) {
-            switch (a.DataType) {
+            switch(a.DataType) {
                 case DataType.Double:
                 case DataType.String:
                 case DataType.Int32:
@@ -317,7 +328,7 @@ namespace AutoJITRuntime
         }
 
         public static Variant operator *( Variant a, Variant b ) {
-            switch (a.DataType) {
+            switch(a.DataType) {
                 case DataType.Int32:
                 case DataType.Int64:
                 case DataType.Double:
@@ -327,30 +338,30 @@ namespace AutoJITRuntime
                     return a.GetInt() * b.GetInt();
                 case DataType.Bool:
                     return a.GetBool()
-                        ? b
-                        : (Variant) 0;
+                    ? b
+                    : (Variant)0;
                 default:
                     throw new NotImplementedException( a.DataType.ToString() );
             }
         }
 
         public static bool operator ==( Variant a, Variant b ) {
-            if ( ReferenceEquals( a, b ) ) {
+            if( ReferenceEquals( a, b ) ) {
                 return true;
             }
 
-            if ( ( (object) a == null ) ||
-                 ( (object) b == null ) ) {
+            if( ( (object)a == null )
+                || ( (object)b == null ) ) {
                 return false;
             }
 
-            if ( a.IsString ||
-                 b.IsString ) {
+            if( a.IsString
+                || b.IsString ) {
                 return string.Compare( a.ToString(), b.ToString(), StringComparison.InvariantCultureIgnoreCase ) == 0;
             }
 
             DataType dataType = GetDatatypeToCompare( a.DataType, b.DataType );
-            switch (dataType) {
+            switch(dataType) {
                 case DataType.Int32:
                     return a.GetInt() == b.GetInt();
                 case DataType.Int64:
@@ -372,7 +383,7 @@ namespace AutoJITRuntime
         }
 
         public static Variant operator ^( Variant a, Variant b ) {
-            switch (a.DataType) {
+            switch(a.DataType) {
                 case DataType.Double:
                 case DataType.Int32:
                     return a.GetInt()^b.GetInt();
@@ -388,7 +399,7 @@ namespace AutoJITRuntime
         }
 
         public static Variant operator |( Variant a, Variant b ) {
-            switch (a.DataType) {
+            switch(a.DataType) {
                 case DataType.Double:
                 case DataType.Int32:
                     return a.GetInt()|b.GetInt();
@@ -415,7 +426,7 @@ namespace AutoJITRuntime
 
         public static Variant operator >( Variant a, Variant b ) {
             DataType toCompare = GetDatatypeToCompare( a.DataType, b.DataType );
-            switch (toCompare) {
+            switch(toCompare) {
                 case DataType.Int32:
                     return a.GetInt() > b.GetInt();
                 case DataType.Int64:
@@ -441,7 +452,7 @@ namespace AutoJITRuntime
 
         public static Variant operator <( Variant a, Variant b ) {
             DataType toCompare = GetDatatypeToCompare( a.DataType, b.DataType );
-            switch (toCompare) {
+            switch(toCompare) {
                 case DataType.Int32:
                     return a.GetInt() < b.GetInt();
                 case DataType.Int64:
@@ -460,9 +471,9 @@ namespace AutoJITRuntime
         private static DataType GetDatatypeToCompare( DataType a, DataType b ) {
             var compareType = DataType.Int32;
 
-            switch (a) {
+            switch(a) {
                 case DataType.Int32:
-                    switch (b) {
+                    switch(b) {
                         case DataType.Int32:
                             compareType = DataType.Int32;
                             break;
@@ -478,7 +489,7 @@ namespace AutoJITRuntime
                     }
                     break;
                 case DataType.Int64:
-                    switch (b) {
+                    switch(b) {
                         case DataType.Int32:
                         case DataType.Int64:
                             compareType = DataType.Int64;
@@ -493,7 +504,7 @@ namespace AutoJITRuntime
                     compareType = DataType.Double;
                     break;
                 case DataType.String:
-                    switch (b) {
+                    switch(b) {
                         case DataType.String:
                             compareType = DataType.String;
                             break;
@@ -513,7 +524,7 @@ namespace AutoJITRuntime
         }
 
         public static implicit operator Variant( Variant[,,] a ) {
-            return new ArrayVariant(a);
+            return new ArrayVariant( a );
         }
 
         public static implicit operator Variant( int a ) {
@@ -550,7 +561,7 @@ namespace AutoJITRuntime
 
         public override bool Equals( object obj ) {
             var variant = obj as Variant;
-            if ( variant != null ) {
+            if( variant != null ) {
                 return GetValue().Equals( variant.GetValue() );
             }
             return GetValue().Equals( obj );
@@ -561,7 +572,7 @@ namespace AutoJITRuntime
         }
 
         public static Variant CreateArray( Variant array ) {
-            Init( (Array) array.GetValue() );
+            Init( (Array)array.GetValue() );
             return array;
         }
 
@@ -571,10 +582,10 @@ namespace AutoJITRuntime
         }
 
         private static void SetDimension( Array array, int[] indicies, int dimension ) {
-            for ( int i = 0; i <= array.GetUpperBound( dimension ); i++ ) {
+            for( int i = 0; i <= array.GetUpperBound( dimension ); i++ ) {
                 indicies[dimension] = i;
 
-                if ( dimension < array.Rank-1 ) {
+                if( dimension < array.Rank-1 ) {
                     SetDimension( array, indicies, dimension+1 );
                 }
                 else {
